+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -48,4 +48,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [NPP-56] 》EX47 - Get center name from school by schoolID
 - [NPP-57] 》EX48 - Update only school name by schoolID
+- [NPP-58] 》EX49 - Add students to a course by DNI and courseID